refactor: extract JavaFile for stub generation
